Rename statistics to metrics
